libcusmm/*.py: unify comments style, add header, refactor
- Merge parameter_space_utils.py with cusmm_dnt_helper.py
- Add folder to hold exploratory notebooks
- Add new header
- Make comment style consistent
